Move home page to pages folder
