+++ src/pages/Home.tsx
@@ -1,29 +1,29 @@
 import React from 'react';
-import './components/Navbar.css';
+import 'components/Navbar.css';
 import './Home.css';
 import { Col, Row } from 'react-flexbox-grid';
-import HomeCard from './components/Card';
-import Storming from './assets/images/Storming.svg';
-import Delivery from './assets/images/Delivery.svg';
-import Augmentation from './assets/images/Augmentation.svg';
-import TypeLogo from './assets/images/TypeLogo.png';
-import Suflex from './assets/images/Suflex.png';
-import Kovi from './assets/images/Kovi.png';
-import Cvc from './assets/images/Cvc.png';
-import Bxblue from './assets/images/Bxblue.png';
-import Bv from './assets/images/Bv.png';
-import Gupy from './assets/images/Gupy.png';
-import FindHotel from './assets/images/FindHotel.png';
-import Poliedro from './assets/images/Poliedro.png';
-import Pipefy from './assets/images/Pipefy.png';
-import Neo from './assets/images/Neo.png';
+import HomeCard from 'components/Card';
+import Storming from 'assets/images/Storming.svg';
+import Delivery from 'assets/images/Delivery.svg';
+import Augmentation from 'assets/images/Augmentation.svg';
+import TypeLogo from 'assets/images/TypeLogo.png';
+import Suflex from 'assets/images/Suflex.png';
+import Kovi from 'assets/images/Kovi.png';
+import Cvc from 'assets/images/Cvc.png';
+import Bxblue from 'assets/images/Bxblue.png';
+import Bv from 'assets/images/Bv.png';
+import Gupy from 'assets/images/Gupy.png';
+import FindHotel from 'assets/images/FindHotel.png';
+import Poliedro from 'assets/images/Poliedro.png';
+import Pipefy from 'assets/images/Pipefy.png';
+import Neo from 'assets/images/Neo.png';
 import {
   HorizontalScroll,
   SliderCard,
   VerticalSeparator,
   HorizontalSeparator,
-} from './components';
-import { Footer } from './components/Footer';
+} from 'components';
+import { Footer } from 'components/Footer';
 import styled from 'styled-components';
 import { Colors } from 'consts/colors';
